Add intermediate code representation and make Toy generator use it.
  Add a file containing the list of available numeric opcodes, and
their corresponding names (and in the future argument lists).
  Adapt the Assembly module to handle opcodes with a parameter
map (and not just a parameter list) and numeric opcodes.
  Move most of the code in the Toy code generator inside a generic
intermediate code generator and make it produce a sequence
of basic blocks instead of a linkear string of code.
  Change all 'op' (for unary/binary operations) and 'function'
(for builtins/function calls) fields to use numeric operation codes
instead of strings.
  For binary opcodes, stack top contains the right operand on
all stack-based virtual machines.